refactor/admin-role-separation - AdminService.grantAdminRole()
